Load all JSON-LD data
The LinkedData class was only grabbing the first application/ld+json
script tag found in the DOM. It now finds them all and makes them
all available via the getAll method. Additionally, it inspects the
mainEntityOfPage property of the json nodes and if a match is found
with the request URI, that node is assigned as the default node
rather than the first node found.